requirements.txt
fastapi[standard]==0.115.12
google-cloud-storage==3.1.0
joblib==1.4.2
keras==3.10.0
matplotlib==3.10.3
nltk==3.9.1
numpy==2.3.0
pandas==2.3.0
protobuf==6.31.1
pydantic==2.11.5
scikit_learn==1.6.1
scipy==1.15.3
shap==0.47.2
tensorflow==2.19.0
uvicorn[standard]==0.34.3
xgboost==3.0.2
sastrawi==1.0.1